requirements.txt
bandcamp-downloader==0.0.13
beautifulsoup4==4.12.2
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
customtkinter==5.1.3
darkdetect==0.8.0
decorator==5.1.1
demjson3==3.0.6
docopt==0.6.2
idna==3.4
imageio==2.28.1
imageio-ffmpeg==0.4.8
lxml==4.9.2
mock==5.0.2
moviepy==1.0.3
mutagen==1.46.0
numpy==1.24.3
Pillow==9.5.0
proglog==0.1.10
pytube==15.0.0
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
tqdm==4.65.0
unicode-slugify==0.1.5
Unidecode==1.3.6
urllib3==2.0.2
